URL names in files  adjusted, without using parcNational

+++ controllers/CompletedTrailsController.php
@@ -5,17 +5,20 @@ require_once __DIR__ . '/../models/CompletedTrails.php';
 
 class CompletedTrailsController extends Controller{
     public function manageCompletedTrail(){
-        $completedTrailsObject = new CompletedTrails('_completed_trails');
-        $trailId = $_GET['trail_id'];
-        $completedTrail = $completedTrailsObject->getCompletedTrail($trailId);
-        if($completedTrail === FALSE){
-            $completedTrailsObject->addCompletedTrail($trailId);
-            $this->redirect('trails');
+        if (isset($_SESSION['user_id'])){
+            $completedTrailsObject = new CompletedTrails('_completed_trails');
+            $trailId = $_GET['trail_id'];
+            $completedTrail = $completedTrailsObject->getCompletedTrail($trailId);
+            if($completedTrail === FALSE){
+                $completedTrailsObject->addCompletedTrail($trailId);
+                $this->redirect('trails');
+            }else{
+                $completedTrailsObject->deleteCompletedTrail($trailId);
+                $this->redirect('profile');
+            }
         }else{
-            $completedTrailsObject->deleteCompletedTrail($trailId);
-            $this->redirect('profile');
+            $this->redirect('login');
         }
-
 
     }
 }
